修复分类保存Bug

+++ publiccms-by-gradle/src/com/publiccms/controller/admin/cms/CmsCategoryAdminController.java
@@ -83,6 +83,7 @@ public class CmsCategoryAdminController extends AbstractController {
      * @return
      */
     @RequestMapping("save")
+    /*Bug：category的保存涉及多张表的操作，应该放入事务中，以便报错时能一起回滚*/
     public String save(CmsCategory entity, CmsCategoryAttribute attribute, @ModelAttribute CmsCategoryParamters categoryParamters,
             HttpServletRequest request, HttpSession session, ModelMap model) {
         SysSite site = getSite(request);
@@ -122,16 +123,21 @@ public class CmsCategoryAdminController extends AbstractController {
         service.updateTagTypeIds(entity.getId(), arrayToCommaDelimitedString(tagTypeIds));// 更新保存标签分类
 
         List<CmsCategoryModelParamters> categoryModelList = categoryParamters.getCategoryModelList();
-        for (CmsCategoryModelParamters cmsCategoryModelParamters : categoryModelList) {
-            if(notEmpty(cmsCategoryModelParamters.getCategoryModel())){
-                cmsCategoryModelParamters.getCategoryModel().getId().setCategoryId(entity.getId());
-                if (cmsCategoryModelParamters.isUse()) {
-                    categoryModelService.updateCategoryModel(cmsCategoryModelParamters.getCategoryModel());
-                } else {
-                    categoryModelService.delete(cmsCategoryModelParamters.getCategoryModel().getId());
+        /* BUG:categoryModelList为空时报java.lang.NullPointerException，需要加验证
+         */
+        if(notEmpty(categoryModelList)){
+        	for (CmsCategoryModelParamters cmsCategoryModelParamters : categoryModelList) {
+                if(notEmpty(cmsCategoryModelParamters.getCategoryModel())){
+                    cmsCategoryModelParamters.getCategoryModel().getId().setCategoryId(entity.getId());
+                    if (cmsCategoryModelParamters.isUse()) {
+                        categoryModelService.updateCategoryModel(cmsCategoryModelParamters.getCategoryModel());
+                    } else {
+                        categoryModelService.delete(cmsCategoryModelParamters.getCategoryModel().getId());
+                    }
                 }
             }
         }
+        
         extendFieldService.update(entity.getExtendId(), categoryParamters.getContentExtends());// 修改或增加内容扩展字段
 
         CmsCategoryType categoryType = categoryTypeService.getEntity(entity.getTypeId());
